refactor(utils): extend existing useStableId with React.useId
Consolidate on the existing src/components/UtilityComponents/useStableId
hook instead of the parallel one introduced in the previous commit. The
existing hook is extended to delegate to React.useId() on React 18+ (so
it is SSR-stable there) while keeping its nanoid fallback for React 17.

- src/utils/useStableId.ts and its test moved into UtilityComponents.
- The four call sites added in the previous commit (ChatView, SwitchField,
  MessageText, SearchBar) now import from UtilityComponents and drop the
  prefix argument.
- ESLint react-compat escape-hatch updated to the new path.
