+++ src/components/UtilityComponents/useStableId.ts
@@ -1,13 +1,17 @@
 import { nanoid } from 'nanoid';
-import { useMemo } from 'react';
+import React, { useMemo } from 'react';
+
+const reactUseId = React.useId as (() => string) | undefined;
 
 /**
- * The ID is generated using the `nanoid` library and is memoized to ensure
- * that it remains the same across renders unless the key changes.
+ * Returns a stable, unique string id.
+ *
+ * On React 18+ this delegates to `React.useId()` (with the surrounding colons stripped
+ * so the value is safe to use anywhere an HTML id is expected) and is SSR-stable.
+ * On React 17, it falls back to a client-only id generated via `nanoid`.
  */
-export const useStableId = (key?: string) => {
-  // eslint-disable-next-line react-hooks/exhaustive-deps
-  const id = useMemo(() => nanoid(), [key]);
-
-  return id;
+export const useStableId = () => {
+  if (reactUseId) return reactUseId().replace(/:/g, '');
+  // eslint-disable-next-line react-hooks/rules-of-hooks
+  return useMemo(() => nanoid(), []);
 };
